+++ src/core/subsystem/systems/SSStartup.h
@@ -6,7 +6,7 @@ public:
 	~SSStartup();
 
 	virtual void Startup();
-	virtual void Update(const float deltaTime);
+	virtual void Update(const double deltaTime);
 	virtual void Shutdown();
 private:
 };
